Fold the menu-item tests into the existing suite; role the anchor
